Виправлення зауважень
