+++ controllers/contacts/getContactById.js
@@ -1,0 +1,16 @@
+const { HttpError } = require("../../helpers");
+const { Contact } = require("../../models/contact");
+
+const getContactById = async (req, res, next) => {
+  const { id } = req.params;
+  const result = await Contact.findById(id);
+  if (!result) {
+    throw HttpError(404, "Not found");
+  }
+  res.json(result);
+};
+
+// module.exports = { getContactById: ctrlWrapper(getContactById) };
+
+
+module.exports = getContactById
