+++ routes/api/contacts.js
@@ -1,6 +1,6 @@
 const express = require("express");
 
-const ctrl = require("../../controllers/contacts");
+const ctrl = require("../../controllers/contacts/contacts");
 const {validateBody, isValidId}= require("../../middlewares");
 const {schemas} = require("../../models/contact")
